lowering case of default verifyText command
It seems as if the IDE is not applying styles to DOM objects when
reading from it, while WebDriver does. The effect is that a string
upcased in CSS will never match both.
Lowering (or upping for that matter) the expected string as well as the
obtained solves this issue, at the cost of less precision, but I can't
think of a better way right now.